requirements.txt
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.13
streamlit>=1.32.0
streamlit-ace>=0.1.1
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
tavily-python>=0.2.2
duckduckgo-search>=4.1.0
playwright>=1.40.0
gunicorn>=21.2.0
